# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Delicious Dinner API",
        "description": "API realizada en Laravel",
        "contact": {
            "email": "ingisraeltrujillo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/estado_domicilio": {
            "get": {
                "tags": [
                    "estado_domicilio"
                ],
                "summary": "Muestra todos los estados domicilio",
                "description": "Muestra todos los estados domicilio.",
                "operationId": "9bf5fe9a94bf225ff34927d2a7e8c424",
                "responses": {
                    "200": {
                        "description": "Muestra todos los estados domicilio"
                    },
                    "500": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "estado_domicilio"
                ],
                "summary": "Crear un nuevo estado domicilio",
                "operationId": "19c3f8164983a496b2af98e71d9d3374",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "codename": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "PENDIENTE",
                                    "codename": "pendiente"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED"
                    }
                }
            }
        },
        "/api/estado_domicilio/{id}": {
            "get": {
                "tags": [
                    "estado_domicilio"
                ],
                "summary": "Mostrar la información de un estado domicilio",
                "operationId": "47cf05aeef6b2b8dc8a3fb4fc3703b78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "estado_domicilio"
                ],
                "summary": "Actualizar un estado domicilio",
                "operationId": "f3d972df3961a8ab5a57e0bec63f3d97",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "codename": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "PENDIENTE",
                                    "codename": "pendiente"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            },
            "delete": {
                "tags": [
                    "estado_domicilio"
                ],
                "summary": "Eliminar un estado domicilio",
                "operationId": "2ab061b0a33cdb617be682504c9dbce9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            }
        },
        "/api/estado_venta": {
            "get": {
                "tags": [
                    "estado_venta"
                ],
                "summary": "Muestra todos los estados venta",
                "description": "Muestra todos los estados venta.",
                "operationId": "58386dc499e1310d07d89469af53e4d6",
                "responses": {
                    "200": {
                        "description": "Muestra todos los estados venta"
                    },
                    "500": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "estado_venta"
                ],
                "summary": "Crear un nuevo estado venta",
                "operationId": "a8e4391e7aa7892fd4a01c04ce40d856",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "codename": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "PENDIENTE",
                                    "codename": "pendiente"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED"
                    }
                }
            }
        },
        "/api/estado_venta/{id}": {
            "get": {
                "tags": [
                    "estado_venta"
                ],
                "summary": "Mostrar la información de un estado venta",
                "operationId": "5e83843205a065e20c9585fb77172051",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "estado_venta"
                ],
                "summary": "Actualizar un estado venta",
                "operationId": "53e9fbd1203fb22153d97a352fc84b90",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nombre": {
                                                "type": "string"
                                            },
                                            "codename": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "PENDIENTE",
                                    "codename": "pendiente"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            },
            "delete": {
                "tags": [
                    "estado_venta"
                ],
                "summary": "Eliminar un estado venta",
                "operationId": "a1279ee57f3405a63be5d08303394a44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            }
        },
...
    }
}